+++ README.md
@@ -13,6 +13,14 @@ git commit -m "your message"
 git push -f origin master
 ```
 
+### Container Presenter Pattern + Index Pattern
+
+- _props_ 는 부모 컴포넌트가 자식 컴포넌트에게 주는 값, 자식 컴포넌트에서는 props 를 받아오기만하고, 받아온 props 를 직접 수정 불가능
+- _state_ 컴포넌트 내부에서 선언하며 내부에서 값을 변경 가능
+- index.js 에서 Container.js Import
+- Container.js : prop, state 를 통해 받은 데이터를 Presenter.js
+- Presenter.js : 넘겨 받은 데이터를 출력
+
 ### 플러그인
 
 - TypeScript React Code Snippets
@@ -75,7 +83,8 @@ git push -f origin master
 ### Plugin
 
 #### [REST Client](https://marketplace.visualstudio.com/items?itemName=humao.rest-client)
-  RESTFUL 미리보기 가능
+
+RESTFUL 미리보기 가능
 
 ```
 GET https://devapi.monpass.im/api/crm/sms/log/?start_date=2019.11.01&end_date=2019.11.30
