+++ src/App.tsx
@@ -3,7 +3,7 @@
  */
 import React from 'react'
 //router
-import Router from 'router'
+import {Router} from 'lib/index'
 //---------------------------------------------------------------------
 function App() {
   return <Router />
