+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+// 메인
+export {default as Main} from './main'
