+++ src/pages/main/container.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import Presenter from './presenter'
+
+//type
+type Props = {}
+
+export default function Container({}: Props) {
+  return <Presenter />
+}
